ETL spark/requirements.txt
pandas
pyarrow
docker
minio
python-dotenv
FlightRadarAPI
apache-airflow